Refactor cctbx interface into multiple files
