+++ pydiscamb/cctbx_interface/gradients.py
@@ -1,71 +1,9 @@
-from typing import Any
-
+from cctbx.array_family import flex
 from cctbx.xray.structure_factors.gradients_base import gradients_base
 from cctbx.xray.structure_factors.gradients_direct import gradients_direct
-from cctbx.xray.structure_factors.manager import managed_calculation_base
-from cctbx.xray.structure_factors.from_scatterers_direct import from_scatterers_direct
-from cctbx.array_family import flex
-import iotbx.phil
-from libtbx.phil import scope_extract
 
+from pydiscamb.cctbx_interface.phil_scope import scope_to_taam_dict
 from pydiscamb.discamb_wrapper import DiscambWrapperCached, FCalcMethod
-from pydiscamb.taam_parameters import get_default_databank
-
-pydiscamb_master_params = iotbx.phil.parse(
-    f"""
-    taam 
-    .help = Transferrable Aspherical Atom Model (TAAM)-specific parameters
-    {{
-        use_mott_bethe = False
-            .type = bool
-            .help = Whether to use Mott-Bethe to convert x-ray scattering factors to electron scattering factors. 
-        unit_cell_charge = 0.0
-            .type = float
-            .help = Unit cell charge, used to scale multipolar parameters
-        scale_pval_to_charge = True
-            .type = bool
-            .help = Whether to scale multipolar parameters to fit the given charge
-        nproc = 1
-            .type = int
-            .help = Number of cores to use for computing
-        scattering_table_untyped = None
-            .type = str
-            .help = Scattering table to use for untyped atoms. Inferred from the structure if not given
-        use_mott_bethe_untyped = False
-            .type = bool
-            .help = Whether to use Mott-Bethe on untyped atoms to convert x-ray scattering factors to electron scattering factors. 
-        freeze_local_coordinate_system = False
-            .type = bool
-            .help = Whether to re-calculate local coordinate systems for all atoms when updating the structure
-        implementation = *standard large_molecule
-            .type = choice
-            .help = Choice of implementation for calculations. "large_molecules" has some optimizations for this usecase
-        bank_path = {get_default_databank()}
-            .type = str
-            .help = Path to TAAM parameter bank file
-            .expert_level = 2
-    }}
-"""
-)
-
-
-def scope_to_taam_dict(scope: scope_extract) -> dict[str, Any]:
-    assert hasattr(scope, "discamb")
-    taam_params = scope.discamb.taam
-    out = {
-        key: getattr(taam_params, key)
-        for key in dir(taam_params)
-        if key[:2] != "__" and getattr(taam_params, key) is not None
-    }
-    # Manually translate params renamed in the scope
-    out["electron_scattering"] = out.pop("use_mott_bethe")
-    out["scale"] = out.pop("scale_pval_to_charge")
-    out["n_cores"] = out.pop("nproc")
-    out["electron_scattering_iam"] = out.pop("use_mott_bethe_untyped")
-    out["frozen_lcs"] = out.pop("freeze_local_coordinate_system")
-    out["algorithm"] = out.pop("implementation")
-
-    return out
 
 
 class gradients_taam(gradients_direct):
@@ -103,30 +41,6 @@ class gradients_taam(gradients_direct):
         )
         self.d_target_d_site_cart_was_used = False
         self.d_target_d_u_cart_was_used = False
-
-
-class from_scatterers_taam(from_scatterers_direct):
-
-    def __init__(
-        self,
-        xray_structure,
-        miller_set,
-        manager=None,
-        algorithm="taam",
-        extra_params=None,
-    ):
-        if extra_params is None:
-            extra_params_dict = {}
-        else:
-            extra_params_dict = scope_to_taam_dict(extra_params)
-
-        # TODO add timings
-        managed_calculation_base.__init__(
-            self, manager, xray_structure, miller_set, algorithm="taam"
-        )
-        self._results = CctbxFromScatterersResult(
-            xray_structure, miller_set, FCalcMethod.TAAM, **extra_params_dict
-        )
 
 
 class CctbxGradientsResult:
@@ -200,13 +114,3 @@ class CctbxGradientsResult:
 
     def packed(self):
         return self._packed
-
-
-class CctbxFromScatterersResult:
-    def __init__(self, xrs, miller_set, method, **kwargs):
-        w = DiscambWrapperCached(xrs, method, **kwargs)
-        w.set_indices(miller_set.indices())
-        self._fcalc = w.f_calc()
-
-    def f_calc(self):
-        return self._fcalc
